added more pattern test cases
